Remove global state from PPCAnalyst.cpp. Little bit of cleanup.
git-svn-id: https://dolphin-emu.googlecode.com/svn/trunk@1580 8ced0084-cf51-0410-be5f-012b33b47a6e

+++ Source/Core/Core/Src/PowerPC/Jit64/Jit_LoadStorePaired.cpp
@@ -182,7 +182,7 @@ void psq_st(UGeckoInstruction inst)
 	{
 		if (gpr.R(a).IsImm() && !update && cpu_info.bSSSE3)
 		{
-			u32 addr = gpr.R(a).offset + offset;
+			u32 addr = (u32)(gpr.R(a).offset + offset);
 			if (addr == 0xCC008000) {
 				// Writing to FIFO. Let's do fast method.
 				CVTPD2PS(XMM0, fpr.R(s));
